Got rid of mlib imdct. You can now use djbfft instead (--enable-djbfft, disabled by default)
liba52's fft implementation is a bit slower than djbfft, but globally
the difference is about 5% on x86. On sparc the difference is larger
though, nearly 30% :/ But, at least djbfft's source code is public, so
sparc people (Hakan) will hopefully be able to figure it out. Or, they
can just link in with djbfft if they want.